Main and Pause Menu
Made a basic main menu as well as a pause menu.

The main menu includes the play button, options button, credits button, and a quit button. The options and credits have nothing in them yet.

The pause menu has a resume button, options button, back to main menu button, and a quit game button. Once again the options has nothing in it yet.